genai/pyproject.toml
[project]
name = "videonotes-genai"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "sse-starlette>=2.1",
  "pydantic>=2.9",
  "httpx>=0.27",
  "python-dotenv>=1.0",
  "langgraph>=0.2.50",
  "langchain-core>=0.3",
  "youtube-transcript-api>=0.6.2,<1.0",
  "yt-dlp>=2024.10.7",
  "anthropic>=0.39",
  "openai>=1.54",
  "google-genai>=0.3",
  "Pillow>=10.4"
]

[project.optional-dependencies]
dev = [
  "ruff>=0.7",
  "pytest>=8.3"
]

[tool.ruff]
line-length = 100
target-version = "py311"
